+++ src/i18n/locales/ja/client-page.json
