+++ src/i18n/locales/ja/footer.json
