+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import acceptLanguage from "accept-language";
-import { fallbackLng, languages, cookieName } from "@/app/i18n";
+import { fallbackLng, languages, cookieName } from "@/i18n";
 
 acceptLanguage.languages(languages);
